requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
openai==1.6.1
apscheduler==3.10.4
Pillow==10.1.0
feedparser==6.0.11
pytz==2023.3
